Safari-611.1.10/LayoutTests/inspector/canvas/recording-2d-memoryLimit-expected.txt

Test that CanvasManager is able to record actions made to 2D canvas contexts.


== Running test suite: Canvas.recording2D
-- Running test case: Canvas.recording2D.memoryLimit
initialState:
  attributes:
    width: 2
    height: 2
  current state:
    setTransform: [[1,0,0,1,0,0]]
    globalAlpha: 1
    globalCompositeOperation: "source-over"
    lineWidth: 1
    lineCap: "butt"
    lineJoin: "miter"
    miterLimit: 10
    shadowOffsetX: 0
    shadowOffsetY: 0
    shadowBlur: 0
    shadowColor: "rgba(0, 0, 0, 0)"
    setLineDash: [[]]
    lineDashOffset: 0
    font: "10px sans-serif"
    textAlign: "start"
    textBaseline: "alphabetic"
    direction: "ltr"
    strokeStyle: "#000000"
    fillStyle: "#000000"
    imageSmoothingEnabled: true
    imageSmoothingQuality: "low"
    setPath: [""]
  parameters:
  content: <filtered>
frames:
  0: (duration) (incomplete)
    0: arc(1, 2, 3, 4, 5, false)
      swizzleTypes: [Number, Number, Number, Number, Number, Boolean]
      trace:
        0: arc
        1: (anonymous function)
        2: ignoreException
        3: (anonymous function)
        4: executeFrameFunction
        5: performActions
        6: Global Code
        7: evaluateWithScopeExtension
        8: (anonymous function)
        9: _wrapCall
